+++ scalding/src/main/scala/commbank/coppersmith/scalding/HiveParquetSink.scala
@@ -22,22 +22,27 @@ import au.com.cba.omnia.maestro.api._
 
 import commbank.coppersmith._, Feature._
 import Partitions.PathComponents
-import FeatureSink.AttemptedWriteToCommitted
+import FeatureSink.{AttemptedWriteToCommitted, MetadataWriter}
 
 /**
   * Parquet FeatureSink implementation - create using HiveParquetSink.apply in companion object.
   */
 case class HiveParquetSink[T <: ThriftStruct : Manifest : FeatureValueEnc, P : TupleSetter] private(
   table:         HiveTable[T, (P, T)],
-  partitionPath: Path
+  partitionPath: Path,
+  metadataWriter: MetadataWriter
 ) extends FeatureSink {
-  def write(features: TypedPipe[(FeatureValue[Value], FeatureTime)]): FeatureSink.WriteResult = {
+  def write(features: TypedPipe[(FeatureValue[Value], FeatureTime)],
+            metadataSet: MetadataSet[Any]): FeatureSink.WriteResult = {
     FeatureSink.isCommitted(partitionPath).flatMap(committed =>
       if (committed) {
         Execution.from(Left(AttemptedWriteToCommitted(partitionPath)))
       } else {
         val eavts = features.map(implicitly[FeatureValueEnc[T]].encode)
-        table.writeExecution(eavts).map(_ => Right(Set(partitionPath)))
+
+        table.writeExecution(eavts).flatMap { _ =>
+          writeMetadata(metadataSet, Set(partitionPath))
+        }
       }
     )
   }
@@ -54,7 +59,8 @@ object HiveParquetSink {
     dbName:    DatabaseName,
     tableName: TableName,
     tablePath: Path,
-    partition: FixedSinkPartition[T, P]
+    partition: FixedSinkPartition[T, P],
+    writeMetadata: MetadataWriter = FeatureSink.defaultMetadataWriter
   ): HiveParquetSink[T, P] = {
     val hiveTable = HiveTable[T, P](
       dbName,
@@ -71,6 +77,6 @@ object HiveParquetSink {
     val pathComponents = implicitly[PathComponents[P]].toComponents(partition.partitionValue)
     val partitionRelPath = new Path(partition.underlying.pattern.format(pathComponents: _*))
 
-    HiveParquetSink[T, P](hiveTable, new Path(tablePath, partitionRelPath))
+    HiveParquetSink[T, P](hiveTable, new Path(tablePath, partitionRelPath), writeMetadata)
   }
 }